requirements.txt
numpy~=1.20.3
seaborn
cmcrameri~=1.2
sklearn~=0.0
pycocotools~=2.0.2
betacal~=1.1.0
uncertainty-calibration
tqdm
dufte~=0.2.27
tikzplotlib~=0.9.12
pandas~=1.3.1
hydra-core
ensemble-boxes
matplotlib~=3.4.2
scikit-learn~=0.24.2
omegaconf~=2.1.1
netcal
pyminizip